Upgrade to tokio 1.0
This commit only focuses on making the lib compatible with tokio 1.0.
There are some breaking interface changes because of how rants exposes
the tokio mpsc::Receiver on subscribe. The tokio v1 interface changes
the interface of receiver so that change is passed on to the end user.
This means we should probably do a minor version bump of the lib should
we merge this patch.

NOTE: Perhaps this commit should be a bit more stict with dependency
versions. I only locked down the major and minor versions. Other
dependencies locked down the patch version as well.